Rethink temporary roots
Automatically add all allocations to the list of temporary roots. This removes the burden of remembering to add allocations to the temporary roots (but only if there will be another allocation before the object is put in the root graph), but means that we have to call `gc_cycle()' more often.