Implement call/cc
I had a call/cc, but for some reason it was a nullary function which makes no sense. I also implemented set! which is necessary for a lot of call/cc algorithms.

llisp can now handle all the examples on the call-with-current-continuation wikipedia page including David Madore's ying-yang puzzle (if you fix the define syntax).