Expand macros at "compile time" instead of "run time"
This isn't 100% solid. I haven't tested the shadowing capabilities with lambda & define. I also haven't tested what happens when you have a defmacro somewhere inside. And it's still possible to pass macros around as arguments to functions, which shouldn't be possible. But it's good enough for the stdlib and simple tests. And it's way, way faster.